Add partial support for triggers on symlinks.
Given two paths to the same event-triggering file (symlink path &
canonical path), only one of the paths was being repoted. Therefore, if
an if-stmt was for the wrong path, nothing would trigger. The change is
that both paths are now tested for triggers.

If there is more than one symlink path, at best, only one of them will
be tested. For this reason, it's recommended that only absolute paths
are used for if-stmts.