Add listen/emit for iffts.
An ifft can now emit a tag after it's completed which another ifft can
listen to and trigger on. This is signaled with an if-condition
formatted as "listen:PATH:TRIGGER".

For the "on start" case, an "on_start_listen:PATH:TRIGGER" if-condition
can be used. A dependency graph is generated to order ifft executions.